+++ .vitepress/theme/components/Page.vue
@@ -51,8 +51,10 @@ defineProps({
 </script>
 
 <style scoped>
+
+
 .post-list {
-    /* border-bottom: 1px dashed var(--vp-c-divider); */
+    /* border-bottom: 1px solid var(--vp-c-divider); */
     padding: 14px 0 14px 0;
 }
 .post-header {